==> Final_Project/2. Final_Project_Google Places API.ipynb <==
This is Ipython Notebook two of five for my final project in General Assembly's data science course.

The following code gathers data from the Google Places API. 

Latitude and longitude coordinates from Craigslist listings are used as inputs to the API to gather relevant 
places of interest near each listing. Gathered places currently include grocery stores and gyms. Code collecting 
movie_theatres, train_stations, airports, subway_stations, Barnes_and_Nobles, Deloitte Offices, and Starbucks 
have been commented out to reduce requests to the the API as these places were not found to provide predictive power.

import requests # Helps construct the request to send to the API
import json # JSON helper functions
from bs4 import BeautifulSoup # Data scraping library
import pandas as pd
import time
import ssl

# Get the data collected from Craigslist and select the location attributes for use in the Google Places API

data = pd.read_csv(r'C:\Users\alsherman\Desktop\GitHub\DataScience_GeneralAssembly\Data\Craigslist_Data_May_3_.csv')
#https://raw.githubusercontent.com/Alexjmsherman/DataScience_GeneralAssembly/master/Data/Craigslist_Data_May_3_.csv
location = zip(data['latitude'],data['longitude'],data['ID'])

# Display data

data.head(3)

# Create a url for each listing with the latitude, longitude, and nearby search terms (e.g. grocery store)

def create_url(loc,types,keyword=''):
    
    url = r'https://maps.googleapis.com/maps/api/place/nearbysearch/json?location=' \
        + str(loc[0]) + ',' + str(loc[1]) + \
        '&radius=1610&rankby=prominence&' \
        + types + keyword + 'key='Add Your Google API Key Here'

    return url

#returns a count of the provided search term (e.g. 20 nearby grocery stores)

def count_nearby_places(url):

    # Make the request
    response = requests.post(url)

    # place data in json object
    json_data = json.loads(response.text)

    # search through results and returns count of term
    place_count = 0
    name = ''
    for ind, search_term in enumerate(json_data['results']):
        place_count = ind + 1
        name = search_term['name'] #if I decide to group by certain names - not currently used
    return place_count

# Create several lists of nearby places by making several search requests to the Google Places API for each listing
# To reduce cals to the API, I have commented out attributes that my analysis revealed as not significant

ID_list = []
grocery_list = []
gym_list = []
#movies_theatre_list = []
#train_station_list = []
#airport_list = []
#subway_station_list = []
#barnes_and_nobles_list = []
#Deloitte_list = []
#Starbucks_list = []

for loc in location:
    #Get the unique id for the listing (use this to combine google places data with listing data)
    ID = loc[2]
    ID_list.append(ID)
    
    #add nearby grocery stores
    types = 'types=grocery_or_supermarket&'
    url = create_url(loc, types)
    grocery_list.append(count_nearby_places(url))
    
    #add nearby gyms
    types = 'types=gym&'
    url = create_url(loc, types)
    gym_list.append(count_nearby_places(url))
    
    #add nearby movie theatres
    #types = 'types=movie_theater&'
    #url = create_url(loc, types)
    #movies_theatre_list.append(count_nearby_places(url))
    
    #add nearby train stations
    #types = 'types=train_station&'
    #url = create_url(loc, types)
    #train_station_list.append(count_nearby_places(url))
  
    #add nearby airports
    #types = 'types=airports&'
    #url = create_url(loc, types)
    #airport_list.append(count_nearby_places(url))
    
    #add nearby subway_station
    #types = 'types=subway_station&'
    #url = create_url(loc, types)
    #subway_station_list.append(count_nearby_places(url))
    
    #add nearby barnes and nobles
    #keyword = 'keyword=barnesandnobles&'
    #url = create_url(loc, keyword)
    #barnes_and_nobles_list.append(count_nearby_places(url))

    #add nearby Deloitte
    #keyword = 'keyword=Deloitte&'
    #url = create_url(loc, keyword)
    #Deloitte_list.append(count_nearby_places(url))
    
    #add nearby Starbucks
    #keyword = 'keyword=Starbucks&'
    #url = create_url(loc, keyword)
    #Starbucks_list.append(count_nearby_places(url))

#zip together all search terms for each listing    
data = zip(ID_list, grocery_list,gym_list)
           #movies_theatre_list,train_station_list,
           #airport_list,subway_station_list,
           #barnes_and_nobles_list,Deloitte_list,Starbucks_list)

data = pd.DataFrame(data)
data.columns = [['ID','grocery_list','gym_list']]
                 #'movie_theatre_list','train_station_list',
                 #'airport_list','subway_station_list',
                 #'Barnes_and_Nobles_list','Deloitte_list','Starbucks_list']]

print data.head()

# Create a csv - this is concatenated with the listing data in a later Ipython Notebook

data.to_csv(r'C:\Users\alsherman\Desktop\GitHub\DataScience_GeneralAssembly\Data\GooglePlacesAPI_May_13_part3.csv', index=False)
