==> Final_Project/3. Craigslist-Data Cleaning.ipynb <==
This is Ipython Notebook three of five for my final project in General Assembly's data science course.

The following code contains contains the data pre-processing: cleaning, transformation, outlier detection, imputation, and feature engineering

In practice, the data cleaning and visualization models were created in parallel. I have chosen to separate them for ease of reproducability.

import pandas as pd
import numpy as np
from dateutil import parser
import matplotlib.pyplot as plt
%matplotlib inline  

# Combine all collected data for data cleaning

# Get the craigslist listing data from the first Ipython Notebook
Listing_data = pd.read_csv(r'C:\Users\alsherman\Desktop\GitHub\DataScience_GeneralAssembly\Data\Craigslist_Data_May_3_.csv')

# Get the Google Places data from the second Ipython Notebook
Google_Places_data = pd.read_csv(r'C:\Users\alsherman\Desktop\GitHub\DataScience_GeneralAssembly\Data\GooglePlacesAPI_May_13.csv')

# Combine both data sets into a single dataframe
data = pd.merge(Listing_data, Google_Places_data, how='inner', on='ID')

# View the Data

print len(data)
data.head(3)

## Data Cleaning

# Clean up the Price Field

try: # The code in the try block works only if listings have '$' in the price field (i.e. the first time the code is run)
    # Determine which listings have a price (i.e. the first character of the title is a dollar sign)
    data['price_to_keep'] = [price[0] for price in data.price]
    data = data[data.price_to_keep == '$']
    data.drop('price_to_keep', axis=1, inplace = True)

    # Remove dollar symbol from price to change string to integer
    data.price = [int(price.replace('$','')) for price in data.price]
except: pass

#Drop all listing that do not have a listed price as they do not apply to my analysis
data = data[data.price > 0]

# Remove outliers with unusually high or low rental prices - listing I would unlikely purchase
data = data[data.price < 2500]
data = data[data.price > 1000]

# Remove listings with price listed as <= $100
#I hypothisize that this is to show up at the top when a user sorts by price
data = data[data.price > 100]

#Clean up bathroom field

data.bathroom.replace(' shared',1, inplace=True)
data.bathroom.replace(' split',1, inplace=True)
data.bathroom.replace(np.nan,0, inplace=True)
data.bathroom = [float(bathroom) for bathroom in data.bathroom]

# Clean up square footage

# Remove listings where square footage was not listed as a number (e.g. square_footage == 'loft')
data.square_footage[data.square_footage == 'loft'] = 0
data = data[data.square_footage != 0]

data.square_footage.fillna(0, inplace=True) #improved square footage imputations completed below in this notebook
data.square_footage = [int(i) for i in data.square_footage]

##Data Transformation

# Transform categorical fields to numeric for scikit-learn 

data.country = np.where(data.country == 'US',1,0)
data.availability = np.where(data.availability == 'available now',1,0)

# Covert animals and smoking variables to numeric (1 = allowed; 0 = not allowed)

data['dog'] = [1 if dog == 'dogs are OK - wooof' or dog == 1 else 0 for dog in data.dog]
data['cat'] = [1 if cat == 'cats are OK - purrr' or cat == 1 else 0 for cat in data.cat]
#smoking is bad metric due to mass of NaN values
data['smoking'] = [0 if smoking == 'no smoking' or smoking == 0 else 1 for smoking in data.smoking]

##Outlier Detection - Duplicate and Fake Listings

#Remove duplicates listings (same listing)
data = data.drop_duplicates()

#Remove duplicate listings (same listing data but different Id) - likely seller repeat posting their listing
data.drop_duplicates(['latitude','longitude','description','image_number'], inplace=True)

#remove data with missing square footage, bedroom, and bathroom (e.g. unrealistic created listings)
#Set ID = 999 and filter out these values
data.ID[(data.bedroom == 0) & (data.bathroom == 0) & (data.square_footage == 0) & (data.image_number == 0)] = 999
data = data[data.ID != 999]

## Impute Missing Values

# Fill in missing values for categorical variables as 'No_Data'
data[data.city == 'nan'] #some rows have text 'nan' instead of special Python charcter
data.city.fillna('No_City_Data', inplace=True)
data.country.fillna('No_Country_Data', inplace=True)
data.state.fillna('No_State_Data', inplace=True)
data.housing_type.fillna('No_Housing_Data', inplace=True)
data.url.fillna('No_Url_Data', inplace=True)
data.description.fillna('No_Description_Data', inplace=True)
data.laundry.fillna('No_Laundry_Data', inplace=True)
data.parking.fillna('No_Parking_Data', inplace=True)
data.availability.fillna('No_Availability_Data', inplace=True)
data.date_available.fillna('No_Date_Availability_Data', inplace=True)

# Fill in missing values for numeric variables as 0
data.latitude.fillna(0, inplace=True)
data.location_data_accuracy.fillna(0, inplace=True)
data.longitude.fillna(0, inplace=True)
data.bathroom.fillna(0, inplace=True)
data.bedroom.fillna(0, inplace=True)
data.cat.fillna(0, inplace=True)
data.dog.fillna(0, inplace=True)
data.image_number.fillna(0, inplace=True)
data.grocery_list.fillna(0, inplace=True)
data.gym_list.fillna(0, inplace=True)
data.smoking.fillna(0, inplace=True)
data.average_image_size.fillna(0, inplace=True) #missing values indicate no images, so values are correct as 0

#Improve Square footage imputations (instead of filling all missing values with 0) 
#by filling in values with the mean of housing type and bedroom groups

#create dataframe with mean square footage for groupings of housing type and bedroom
data_square_footage_mean = data.groupby(['housing_type','bedroom'])['square_footage'].mean()
data.square_footage_mean = pd.DataFrame(data.square_footage)
data_square_footage_mean = data_square_footage_mean.reset_index()

#create dataframe with count of groupings of housing type and bedroom
data_square_footage_count = data.groupby(['housing_type','bedroom'])['square_footage'].count()
data.square_footage_count = pd.DataFrame(data.square_footage)
data_square_footage_count = data_square_footage_count.reset_index()

#Combine mean and count groups from above - use this to detect outliers to not use for square footage imputations
#e.g. square footage below 100 or potentially groups with only a few listings
data_square_footage = pd.merge(data_square_footage_mean,data_square_footage_count,how='inner',on=['housing_type','bedroom'])
data_square_footage.columns = [['housing_type','bedroom','square_footage','count']]
#For obviously skewed values from groupby (due to lack of housing type/bedroom group data) replace with mean value
data_square_footage.square_footage[data_square_footage.square_footage < 100] = data_square_footage.square_footage.mean()

#create a dataframe merging the original data and the new data with every value of square footage equaling the mean
xyz = pd.merge(data,data_square_footage, on=['housing_type','bedroom'],how='left')
xyz[['housing_type','bedroom','square_footage_x','square_footage_y']] #reduce to only relevant fields

#create a list using mean square footage values when the original value == 0 (NaN before replacement)
#else use the original values
square_footage = []
for ind, i in enumerate(xyz.square_footage_x):
    if xyz.square_footage_x[ind] == 0:
        square_footage.append(int(xyz.square_footage_y[ind]))
    else: 
        square_footage.append(int(xyz.square_footage_x[ind]))

#Add clean square_footage to data
data['square_footage_cleaned'] = square_footage      


## Feature Engineering

# Split time and date of posting into separate columns

data['time_of_day_of_posting'] = [time.split()[1] for time in data.time_of_posting]
data['date_of_posting'] = [time.split()[0] for time in data.time_of_posting]
data['weekday_of_posting'] = [parser.parse(day).weekday() for day in data.date_of_posting]

#fill in missing values
data.time_of_day_of_posting.fillna('No_Time_of_Posting_Data', inplace=True)
data.date_of_posting.fillna('No_Date_of_Posting_Data', inplace=True)

# Drop original time_of_posting field with combined date and time
data.drop('time_of_posting', axis=1, inplace = True)

# Turn posting dates and times into numeric data

data['Posting_Time_AM_or_PM'] = [Time[-2:] for Time in data.time_of_day_of_posting]
data['Posting_Time_AM_or_PM'] = [int(Time.replace('am','0').replace('pm','1')) for Time in data.Posting_Time_AM_or_PM]
data['Posting_Time'] = [float(Time.replace('am','').replace('pm','').replace(':','.')) for Time in data.time_of_day_of_posting]
data['Posting_Day'] = [int(Date.split('-')[2]) for Date in data.date_of_posting]

##Create CSV of Cleaned Data for Modeling

data.to_csv(r'C:\Users\alsherman\Desktop\GitHub\DataScience_GeneralAssembly\Data\Cleaned_Data_Final.csv', index=False)

##Depreciated Code

#Remove null values and correlated values

# Airport field consists of only 0 (no listing is near an airport according to Google Places API data)
# Remove this variable as it provides no value
#data.drop('airport_list', axis=1, inplace=True)

# Drop train_station_list as it is correlated with subway_station_list
#data.drop('train_station_list', axis=1, inplace=True)

# Clean up mispelled cities and combine similarly spelled cities

# Turn all cities to lower case
data.city = [str(city).lower() for city in data.city]

# Replace encodings with empty strings
data.city = [str(city).replace('%2c','') for city in data.city]
data.city = [str(city).replace('%2d','') for city in data.city]
data.city = [str(city).replace('%2e','') for city in data.city]
data.city = [str(city).replace('%2f','') for city in data.city]
data.city = [str(city).replace('%21','') for city in data.city]

# Correct cities
data.city[data.city == 'aleandria'] = 'Alexandria'
data.city[data.city == 'anandale'] = 'annandale'
data.city[data.city == 'arliington'] = 'arlington'
data.city[data.city == 'arlilngton'] = 'arlington'
data.city[data.city == 'aspin'] = 'aspen'
data.city[data.city == 'betheda'] = 'bethesda'
data.city[data.city == 'hyattsvile'] = 'hyattsville'
data.city[data.city == 'hyattsvelle'] = 'hyattsville'
data.city[data.city == 'hyatsville'] = 'hyattsville'
data.city[data.city == 'w'] = 'washingtondc'
data.city[data.city == 'wahsington'] = 'washingtondc'
data.city[data.city == 'wash'] = 'washingtondc'
data.city[data.city == 'washigton'] = 'washingtondc'
data.city[data.city == 'washingon'] = 'washingtondc'
data.city[data.city == 'washinnton'] = 'washingtondc'
data.city[data.city == 'washinton'] = 'washingtondc'
data.city[data.city == 'washiongton'] = 'washingtondc'
data.city[data.city == 'dc'] = 'washingtondc'
data.city[data.city == 'district'] = 'washingtondc'
data.city[data.city == 'washington'] = 'washingtondc'
data.city[data.city == 'woodbrige'] = 'woodbridge'
data.city[data.city ==  'forstville'] = 'forestville'
data.city[data.city ==  'gathersburg'] = 'gaithersburg'
data.city[data.city ==  'germatnown'] = 'germantown'
data.city[data.city == 'springfiel'] = 'springfield'
data.city[data.city == 'rockvillenorth'] = 'rockville'
data.city[data.city == 'nw'] = 'northwest'
data.city[data.city == 'norht'] = 'north'
data.city[data.city == 'mt'] = 'mount'

# Clean up mispelled states and combine similarly spelled states

# Turn all cities to lower case
data.state = [str(state).lower() for state in data.state]

# Replace encodings with empty strings
data.state = [str(state).replace('%2e','') for state in data.state]
data.state = [str(state).replace('%2ec','') for state in data.state]
data.state = [str(state).replace('%2f','') for state in data.state]

# Correct states
data.state[data.state == 'arlington'] = 'State_Arlington' #to avoid both a city and state called arlington
data.state[data.state == 'bethesdarockville'] = 'bethesda'
data.state[data.state == 'bethsda'] = 'bethesda'
data.state[data.state == 'd'] = 'dc'
