==> Final_Project/5. Final_Project_Data_Modeling.ipynb <==
This is Ipython Notebook five of five for my final project in General Assembly's data science course.

The following code contains contains the data modeling.

# imports
import pandas as pd
import numpy as np
import statsmodels.formula.api as smf
from sklearn.linear_model import LinearRegression
from sklearn import metrics
from sklearn.cross_validation import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import Ridge
from sklearn.linear_model import RidgeCV 
from sklearn.naive_bayes import MultinomialNB
from sklearn.feature_extraction.text import CountVectorizer

# allow plots to appear directly in the notebook
import seaborn as sns
%matplotlib inline

data = pd.read_csv(r'C:\Users\alsherman\Desktop\GitHub\DataScience_GeneralAssembly\Data\Cleaned_Data_Final.csv')

data.describe()

# Create dummy variables

data = pd.concat([data, pd.get_dummies(data.housing_type)], axis=1)
data = pd.concat([data, pd.get_dummies(data.laundry)], axis=1)
data = pd.concat([data, pd.get_dummies(data.parking)], axis=1)

data[0:1]

feature_cols = [
'latitude',
'longitude',
'bathroom',
'bedroom',
'image_number',
#'square_footage', 
'square_footage_cleaned',
'grocery_list',
'gym_list',  
'attached garage',
'off-street parking',
'street parking',     
'Posting_Time_AM_or_PM',
'Posting_Time',
'Posting_Day',
'weekday_of_posting'
]

## Train Test Split

# create X and y
X = data[feature_cols]
y = data.price

# Train test split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 1)

##Random Forest

# import class, instantiate estimator, fit with all data
rfclf = RandomForestClassifier(n_estimators=200, max_features=5, oob_score=True, random_state=1)
rfclf.fit(X_train, y_train)

# compute the out-of-bag classification accuracy
print rfclf.oob_score_
print len(feature_cols)
print len(rfclf.feature_importances_)

# compute the feature importances
Feature_Importance = pd.DataFrame({'feature':feature_cols, 'importance':rfclf.feature_importances_})
Feature_Importance.sort('importance', ascending=False)

## Linear Regression

### SCIKIT-LEARN ###

# instantiate and fit
lm2 = LinearRegression()
lm2.fit(X_train, y_train)

y_pred = lm2.predict(X_test) # Predict
# print the intercept
#print lm2.intercept_

# pair the feature names with the coefficients
#print zip(feature_cols, lm2.coef_)

# Access accuracy
y_pred = [int(y) for y in y_pred]
y_test = [int(y) for y in y_test]

print metrics.accuracy_score(y_test, y_pred)

##Ridge Regression

rreg = Ridge(alpha=0.01, normalize=True) 
rreg.fit(X_train, y_train) 
rreg.coef_ 
preds = rreg.predict(X_test) 
print 'RMSE (Ridge reg.) =', np.sqrt(metrics.mean_squared_error(y_test, preds)) 

# Access accuracy
preds = [int(y) for y in preds]
y_test = [int(y) for y in y_test]
print metrics.accuracy_score(y_test, preds)

alpha_range = 10.**np.arange(-2, 3) 
rregcv = RidgeCV(normalize=True, scoring='mean_squared_error', alphas=alpha_range) 
rregcv.fit(X_train, y_train) 
print rregcv.alpha_ 
preds = rregcv.predict(X_test) 
print 'RMSE (Ridge CV reg.) =', np.sqrt(metrics.mean_squared_error(y_test, preds)) 

##Lasso Regression

from sklearn.linear_model import Lasso 
las = Lasso(alpha=0.01, normalize=True) 
las.fit(X_train, y_train) 
las.coef_ 
preds = las.predict(X_test) 
print 'RMSE (Lasso reg.) =', np.sqrt(metrics.mean_squared_error(y_test, preds))  
  
from sklearn.linear_model import LassoCV 
lascv = LassoCV(normalize=True, alphas=alpha_range) 
lascv.fit(X_train, y_train) 
print lascv.alpha_ 
print zip(feature_cols,lascv.coef_) 
preds = lascv.predict(X_test) 
print 'RMSE (Lasso CV reg.) =', np.sqrt(metrics.mean_squared_error(y_test, preds)) 

# Access accuracy
preds = [int(y) for y in preds]
y_test = [int(y) for y in y_test]
print metrics.accuracy_score(y_test, preds)

## Depreciated Models

##Naive Bayes

 #saving for future use case

#Create Train Test split for text data (listing description)

naive_bayes_feature_cols = 'description'

# create X and y
X = data[naive_bayes_feature_cols]
#y = data

# Train test split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 1)

# COUNTVECTORIZER: 'convert text into a matrix of token counts'
# instantiate the vectorizer
vect = CountVectorizer()

# learn vocabulary and create document-term matrix in a single step
train_dtm = vect.fit_transform(X_train)
train_dtm

# transform testing data into a document-term matrix
test_dtm = vect.transform(X_test)
test_dtm

# store feature names and examine them
train_features = vect.get_feature_names()
len(train_features)
train_features[:50]
train_features[-50:]

# convert train_dtm to a regular array
train_arr = train_dtm.toarray()
train_arr

# examine the vocabulary and document-term matrix together
pd.DataFrame(train_dtm.toarray(), columns=vect.get_feature_names()).head()

nb = MultinomialNB()
nb.fit(train_dtm, y_train)

# make predictions on test data using test_dtm
y_pred = nb.predict(test_dtm)
y_pred

# compare predictions to true labels
print metrics.accuracy_score(y_test, y_pred)
#print metrics.confusion_matrix(y_test, y_pred)
